myapp/chatbot/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.3
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Frontend
streamlit==1.39.0

# LangChain Ecosystem
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-postgres==0.0.16
langchain-neo4j==0.2.0
langgraph==0.2.45
langgraph-checkpoint==2.0.2

# OpenAI
openai==1.54.3

# Database Drivers
neo4j==5.25.0
psycopg2-binary==2.9.3
supabase==2.22.0

# HTTP Client
httpx==0.27.2

# Utilities
tiktoken==0.8.0
tenacity==9.0.0

# Optional
langsmith==0.1.147